+++ test/translate/__init__.py
@@ -1,0 +1,4 @@
+"""Tests for the minmax normalizer
+
+@since 0.7.0
+"""