+++ tf_rnn/translate/__init__.py
@@ -1,0 +1,9 @@
+"""Classes for translating inputs and outputs from the RNN format to a human-understandable format.
+
+@since 0.7.0
+"""
+
+from .translators import Translators
+from .translator import Translator
+from .tokenizer import Tokenizer
+from .minmax_normalizer import MinMaxNormalizer